requirements.txt
aiogram==3.10.0
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
loguru==0.7.2
gspread==6.1.2
google-auth==2.35.0
